services/api/requirements.txt
# Core Framework
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.3
pydantic-settings==2.2.1

# Database
sqlalchemy==2.0.28
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Cache & Queue
redis==5.0.2

# Object Storage
minio==7.2.5

# Auth
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
httpx==0.27.0

# Observability
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-instrumentation-sqlalchemy==0.44b0
opentelemetry-instrumentation-redis==0.44b0
opentelemetry-exporter-otlp==1.23.0
prometheus-client==0.20.0
prometheus-fastapi-instrumentator==7.0.0

# Utilities
python-dotenv==1.0.1
typer==0.9.0
rich==13.7.1
structlog==24.1.0
